# backend/internal/server/openapi.yaml
openapi: 3.1.0
info:
  title: Mansariya API
  description: |
    Sri Lanka's first crowdsource-powered bus tracking API.
    Passengers' phone GPS becomes the sensor — no hardware on buses.

    ## Authentication
    - **Public endpoints** require no authentication.
    - **Admin endpoints** require an `X-API-Key` header.

    ## Real-time Updates
    Live bus positions are streamed via WebSocket at `ws://host/ws/track/{routeID}`.

    ## Trilingual Support
    Route and stop names are available in English (`name_en`), Sinhala (`name_si`), and Tamil (`name_ta`).
  version: 1.0.0
  contact:
    name: Mansariya Team
    url: https://github.com/theetaz/mansariya
  license:
    name: MIT
servers:
  - url: https://api.masariya.lk
    description: Production
  - url: http://localhost:9900
    description: Local development
...
paths:
...
  /api/v1/gps/batch:
    post:
      tags: [GPS]
      summary: Submit GPS batch
      description: |
        Receives GPS observations from a mobile device and pushes them into the
        processing pipeline via Redis Stream. Returns immediately — processing is async.
      operationId: gpsBatch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GPSBatch"
            example:
              device_hash: "a1b2c3d4e5f6"
              session_id: "sess_20240327_001"
              pings:
                - lat: 6.9271
                  lng: 79.8612
                  ts: 1711540800
                  acc: 12.5
                  spd: 8.3
                  brg: 45.0
      responses:
        "200":
          description: Batch accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  processed:
                    type: integer
                    example: 10
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/routes:
    get:
      tags: [Routes]
      summary: List routes near location
      operationId: listRoutes
      parameters:
        - name: lat
          in: query
          required: true
          schema:
            type: number
            format: double
          example: 6.9271
        - name: lng
          in: query
          required: true
          schema:
            type: number
            format: double
          example: 79.8612
        - name: radius_km
          in: query
          schema:
            type: number
            format: double
            default: 5
          description: Search radius in kilometers
      responses:
        "200":
          description: List of nearby routes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Route"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/routes/sync:
    get:
      tags: [Sync]
      summary: Sync all routes for offline caching
      description: |
        Returns all active routes. Called by mobile on first launch and weekly
        to populate the local SQLite cache.
      operationId: syncRoutes
      responses:
        "200":
          description: All active routes
          content:
            application/json:
              schema:
                type: object
                properties:
                  routes:
                    type: array
                    items:
                      $ref: "#/components/schemas/Route"
                  count:
                    type: integer
                    example: 15
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/routes/{routeID}:
    get:
      tags: [Routes]
      summary: Get route with stops
      operationId: getRoute
      parameters:
        - $ref: "#/components/parameters/RouteID"
      responses:
        "200":
          description: Route detail with stops
          content:
            application/json:
              schema:
                type: object
                properties:
                  route:
                    $ref: "#/components/schemas/Route"
                  stops:
                    type: array
                    items:
                      $ref: "#/components/schemas/Stop"
                    nullable: true
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/routes/{routeID}/polyline:
    get:
      tags: [Routes]
      summary: Get route polyline geometry
      operationId: getRoutePolyline
      parameters:
        - $ref: "#/components/parameters/RouteID"
      responses:
        "200":
          description: Road-snapped route geometry
          content:
            application/json:
              schema:
                type: object
                properties:
                  route_id:
                    type: string
                    example: "138"
                  coordinates:
                    type: array
                    items:
                      type: array
                      items:
                        type: number
                      minItems: 2
                      maxItems: 2
                    description: "[lng, lat] coordinate pairs"
                  points:
                    type: integer
                    example: 245
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/routes/{routeID}/eta:
    get:
      tags: [ETA]
      summary: Get bus ETAs for a stop on a route
      description: |
        Returns estimated arrival times for all active buses heading toward
        the specified stop location on the given route.
      operationId: getETA
      parameters:
        - $ref: "#/components/parameters/RouteID"
        - name: lat
          in: query
          required: true
          schema:
            type: number
            format: double
          example: 6.9271
          description: Stop latitude
        - name: lng
          in: query
          required: true
          schema:
            type: number
            format: double
          example: 79.8612
          description: Stop longitude
      responses:
        "200":
          description: ETA estimates for approaching buses
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ETAResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/routes/{routeID}/stops:
    get:
      tags: [Routes]
      summary: Get enriched stops for a route
      description: Returns stops with ordering, timing, and fare data.
      operationId: getRouteStops
      parameters:
        - $ref: "#/components/parameters/RouteID"
      responses:
        "200":
          description: Ordered list of enriched route stops
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EnrichedRouteStop"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/search:
    get:
      tags: [Search]
      summary: Search routes (trilingual)
      description: |
        Full-text search across route names in English, Sinhala, and Tamil.
        Uses PostgreSQL `pg_trgm` for fuzzy matching.
      operationId: searchRoutes
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
          example: "Colombo Kandy"
          description: Search query in any language
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 50
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  query:
                    type: string
                  results:
                    type: array
                    items:
                      $ref: "#/components/schemas/Route"
                  count:
                    type: integer
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/journey:
    get:
      tags: [Journey]
      summary: Plan a journey between two stops
      description: |
        Finds all routes that connect two stops, with estimated duration,
        fares, and live bus information.
      operationId: searchJourney
      parameters:
        - name: from
          in: query
          required: true
          schema:
            type: string
          example: "Kadawatha"
          description: Origin stop name
        - name: to
          in: query
          required: true
          schema:
            type: string
          example: "Kegalle"
          description: Destination stop name
      responses:
        "200":
          description: Journey options
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JourneyResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/stops/nearby:
    get:
      tags: [Stops]
      summary: Find stops near a location
      operationId: nearbyStops
      parameters:
        - name: lat
          in: query
          required: true
          schema:
            type: number
            format: double
          example: 6.9271
        - name: lng
          in: query
          required: true
          schema:
            type: number
            format: double
          example: 79.8612
        - name: radius_km
          in: query
          schema:
            type: number
            format: double
            default: 1
      responses:
        "200":
          description: Nearby stops
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Stop"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/stops/search:
    get:
      tags: [Stops]
      summary: Search stops by name
      description: For autocomplete — returns matching stops across all languages.
      operationId: searchStops
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
          example: "kadaw"
      responses:
        "200":
          description: Matching stops
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Stop"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/buses/active:
    get:
      tags: [Buses]
      summary: Get all active buses
      description: Returns all live bus positions currently tracked in Redis.
      operationId: activeBuses
      responses:
        "200":
          description: Active bus positions
          content:
            application/json:
              schema:
                type: object
                properties:
                  buses:
                    type: array
                    items:
                      $ref: "#/components/schemas/Vehicle"
                  count:
                    type: integer
                    example: 12
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/buses/nearby:
    get:
      tags: [Buses]
      summary: Get buses near a location
      operationId: nearbyBuses
      parameters:
        - name: lat
          in: query
          required: true
          schema:
            type: number
            format: double
          example: 6.9271
        - name: lng
          in: query
          required: true
          schema:
            type: number
            format: double
          example: 79.8612
        - name: radius_km
          in: query
          schema:
            type: number
            format: double
            default: 10
      responses:
        "200":
          description: Buses within radius
          content:
            application/json:
              schema:
                type: object
                properties:
                  buses:
                    type: array
                    items:
                      $ref: "#/components/schemas/Vehicle"
                  count:
                    type: integer
                  radius_km:
                    type: number
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /ws/track/{routeID}:
    get:
      tags: [WebSocket]
      summary: Stream live bus positions (WebSocket)
      description: |
        Upgrades to WebSocket and streams real-time bus position updates for a route.

        **Protocol:** `ws://` or `wss://`

        **Server → Client messages:**
        ```json
        {
          "virtual_id": "v_001",
          "route_id": "138",
          "lat": 6.935,
          "lng": 79.875,
          "speed_kmh": 45.2,
          "bearing": 120.5,
          "contributor_count": 3,
          "confidence": "verified",
          "last_update": "2024-03-27T14:22:30Z"
        }
        ```
      operationId: wsTrack
      parameters:
        - $ref: "#/components/parameters/RouteID"
      responses:
        "101":
          description: Switching protocols (WebSocket upgrade)
        "400":
          $ref: "#/components/responses/BadRequest"

  # ── Admin Endpoints ──
  /api/v1/admin/routes:
    post:
      tags: [Admin Routes]
      summary: Create a route
      operationId: adminCreateRoute
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminRouteInput"
      responses:
        "201":
          description: Route created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/admin/routes/{routeID}:
    put:
      tags: [Admin Routes]
      summary: Update a route
      operationId: adminUpdateRoute
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/RouteID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminRouteInput"
      responses:
        "200":
          description: Route updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      tags: [Admin Routes]
      summary: Delete a route
      operationId: adminDeleteRoute
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/RouteID"
      responses:
        "200":
          description: Route deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/admin/routes/{routeID}/validate:
    post:
      tags: [Admin Routes]
      summary: Mark a route as validated
      operationId: adminValidateRoute
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/RouteID"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                validated_by:
                  type: string
                  example: admin
      responses:
        "200":
          description: Route validated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/admin/routes/{routeID}/stops:
    put:
      tags: [Admin Routes]
      summary: Set stops for a route
      operationId: adminSetRouteStops
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/RouteID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                stops:
                  type: array
                  items:
                    $ref: "#/components/schemas/AdminRouteStopInput"
      responses:
        "200":
          description: Route stops updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  route_id:
                    type: string
                  stops:
                    type: integer
                  status:
                    type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/admin/routes/{routeID}/timetable:
    put:
      tags: [Admin Routes]
      summary: Set timetable for a route
      description: Replaces the existing timetable entries for a route.
      operationId: adminSetTimetable
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/RouteID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                entries:
                  type: array
                  items:
                    $ref: "#/components/schemas/AdminTimetableInput"
      responses:
        "200":
          description: Timetable updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  route_id:
                    type: string
                  entries:
                    type: integer
                  status:
                    type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/admin/stops:
    post:
      tags: [Admin Stops]
      summary: Create a stop
      operationId: adminCreateStop
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminStopInput"
      responses:
        "201":
          description: Stop created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/admin/stops/{stopID}:
    put:
      tags: [Admin Stops]
      summary: Update a stop
      operationId: adminUpdateStop
      security:
        - ApiKeyAuth: []
      parameters:
        - name: stopID
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminStopInput"
      responses:
        "200":
          description: Stop updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Admin API key (set via `ADMIN_API_KEY` env var)
  parameters:
    RouteID:
      name: routeID
      in: path
      required: true
      schema:
        type: string
      example: "138"
      description: Route identifier
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Invalid or missing API key
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: "descriptive error message"
      required: [error]
    StatusResponse:
      type: object
      properties:
        id:
          type: string
        status:
          type: string
          enum: [created, updated, deleted, validated]
    GPSBatch:
      type: object
      required: [device_hash, session_id, pings]
      properties:
        device_hash:
          type: string
          description: One-way hash of device UUID (rotates every 24h)
          example: "a1b2c3d4e5f6"
        session_id:
          type: string
          description: Unique tracking session identifier
          example: "sess_20240327_001"
        pings:
          type: array
          items:
            $ref: "#/components/schemas/GPSPing"
          minItems: 1
    GPSPing:
      type: object
      properties:
        lat:
          type: number
          format: double
          description: WGS84 latitude
          example: 6.9271
        lng:
          type: number
          format: double
          description: WGS84 longitude
          example: 79.8612
        ts:
          type: integer
          format: int64
          description: Unix timestamp (seconds)
          example: 1711540800
        acc:
          type: number
          format: double
          description: GPS accuracy in meters
          example: 12.5
        spd:
          type: number
          format: double
          description: Speed in m/s
          example: 8.3
        brg:
          type: number
          format: double
          description: Bearing in degrees (0-360)
          example: 45.0
    Route:
      type: object
      properties:
        id:
          type: string
          example: "138"
        name_en:
          type: string
          example: "Colombo - Kandy"
        name_si:
          type: string
          example: "කොළඹ - මහනුවර"
        name_ta:
          type: string
          example: "கொழும்பு - கண்டி"
        operator:
          type: string
          example: "SLTB"
        service_type:
          type: string
          example: "intercity"
        fare_lkr:
          type: integer
          example: 350
        frequency_minutes:
          type: integer
          example: 15
        operating_hours:
          type: string
          example: "05:00-22:00"
        polyline:
          type: array
          items:
            type: array
            items:
              type: number
          description: "[lng, lat] coordinate pairs"
        polyline_confidence:
          type: number
          format: double
          example: 0.95
        source:
          type: string
          example: "gps_traces"
        is_active:
          type: boolean
          example: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Stop:
      type: object
      properties:
        id:
          type: string
          example: "138_s0"
        name_en:
          type: string
          example: "Colombo Fort"
        name_si:
          type: string
          example: "කොළඹ කොටුව"
        name_ta:
          type: string
          example: "கொழும்பு கோட்டை"
        location:
          type: array
          items:
            type: number
          minItems: 2
          maxItems: 2
          description: "[lng, lat]"
          example: [79.8612, 6.9271]
        source:
          type: string
          example: "bootstrap"
        confidence:
          type: number
          format: double
          example: 0.92
        observation_count:
          type: integer
          example: 450
        created_at:
          type: string
          format: date-time
    Vehicle:
      type: object
      properties:
        virtual_id:
          type: string
          description: Fused virtual bus identifier
          example: "v_001"
        route_id:
          type: string
          example: "138"
        lat:
          type: number
          format: double
          example: 6.935
        lng:
          type: number
          format: double
          example: 79.875
        speed_kmh:
          type: number
          format: double
          example: 45.2
        bearing:
          type: number
          format: double
          example: 120.5
        contributor_count:
          type: integer
          description: Number of phones contributing GPS for this bus
          example: 3
        confidence:
          type: string
          enum: [low, good, verified]
          description: "low = 1 contributor, good = 2, verified = 3+"
          example: "verified"
        last_update:
          type: string
          format: date-time
    ETAResponse:
      type: object
      properties:
        route_id:
          type: string
          example: "138"
        stop_id:
          type: string
          example: "138_s5"
        buses:
          type: array
          items:
            $ref: "#/components/schemas/BusETA"
    BusETA:
      type: object
      properties:
        bus_id:
          type: string
          example: "v_001"
        eta_minutes:
          type: integer
          description: Minutes until bus arrives at stop
          example: 8
        confidence:
          type: string
          enum: [low, good, verified]
          example: "verified"
        distance_km:
          type: number
          format: double
          example: 2.5
        contributor_count:
          type: integer
          example: 3
    EnrichedRouteStop:
      type: object
      properties:
        route_id:
          type: string
          example: "138"
        stop_id:
          type: string
          example: "138_s0"
        stop_order:
          type: integer
          example: 1
        distance_from_start_km:
          type: number
          format: double
          example: 0
        typical_duration_min:
          type: integer
          example: 0
        fare_from_start_lkr:
          type: integer
          example: 0
        stop_name_en:
          type: string
          example: "Colombo Fort"
        stop_name_si:
          type: string
        stop_name_ta:
          type: string
        stop_lat:
          type: number
          format: double
          example: 6.9271
        stop_lng:
          type: number
          format: double
          example: 79.8612
        is_terminal:
          type: boolean
          example: true
    JourneyResponse:
      type: object
      properties:
        origin:
          $ref: "#/components/schemas/Stop"
        destination:
          $ref: "#/components/schemas/Stop"
        journeys:
          type: array
          items:
            $ref: "#/components/schemas/JourneyResult"
    JourneyResult:
      type: object
      properties:
        route:
          $ref: "#/components/schemas/Route"
        board_stop:
          $ref: "#/components/schemas/EnrichedRouteStop"
        exit_stop:
          $ref: "#/components/schemas/EnrichedRouteStop"
        stops_between:
          type: integer
          example: 13
        estimated_duration_min:
          type: integer
          example: 45
        fare_lkr:
          type: integer
          example: 350
        live_bus_count:
          type: integer
          example: 2
        next_bus_eta_min:
          type: integer
          example: 8
    AdminRouteInput:
      type: object
      required: [id, name_en]
      properties:
        id:
          type: string
          example: "r_new"
        name_en:
          type: string
          example: "Colombo - Galle"
        name_si:
          type: string
        name_ta:
          type: string
        operator:
          type: string
          example: "SLTB"
        service_type:
          type: string
          example: "intercity"
        fare_lkr:
          type: integer
          example: 450
        frequency_minutes:
          type: integer
          example: 20
        operating_hours:
          type: string
          example: "05:00-22:00"
        data_source:
          type: string
          example: "manual_entry"
    AdminStopInput:
      type: object
      required: [name_en, lat, lng]
      properties:
        id:
          type: string
        name_en:
          type: string
          example: "Galle Fort"
        name_si:
          type: string
        name_ta:
          type: string
        lat:
          type: number
          format: double
          example: 6.0329
        lng:
          type: number
          format: double
          example: 80.2168
        road_name:
          type: string
        landmark:
          type: string
        is_terminal:
          type: boolean
    AdminRouteStopInput:
      type: object
      properties:
        stop_id:
          type: string
          example: "s001"
        stop_order:
          type: integer
          example: 1
        distance_from_start_km:
          type: number
          format: double
          example: 0
        typical_duration_min:
          type: integer
          example: 0
        fare_from_start_lkr:
          type: integer
          example: 0
    AdminTimetableInput:
      type: object
      properties:
        departure_time:
          type: string
          description: "HH:MM format"
          example: "05:30"
        days:
          type: array
          items:
            type: string
          example: ["MON", "TUE", "WED", "THU", "FRI"]
        service_type:
          type: string
          example: "peak"
        notes:
          type: string
          example: "Morning express service"